refaktorering - flyttet hjelpekode for karakterutskriftbrev til egen util fil
