+++ src/test/kotlin/no/nav/familie/ef/iverksett/brev/frittstående/DistribuerKarakterutskriftBrevTaskTest.kt
@@ -6,7 +6,7 @@ import io.mockk.slot
 import io.mockk.verify
 import no.nav.familie.ef.iverksett.brev.DistribuerJournalpostResponseTo
 import no.nav.familie.ef.iverksett.brev.JournalpostClient
-import no.nav.familie.ef.iverksett.oppgave.OppgaveUtil.opprettBrev
+import no.nav.familie.ef.iverksett.brev.frittstående.KarakterInnhentingBrevUtil.opprettBrev
 import no.nav.familie.ef.iverksett.repository.findByIdOrThrow
 import no.nav.familie.ef.iverksett.vedtakstatistikk.toJson
 import no.nav.familie.http.client.RessursException
