+++ src/test/kotlin/no/nav/familie/ef/iverksett/brev/frittstående/JournalførKarakterutskriftBrevTaskTest.kt
@@ -6,7 +6,7 @@ import io.mockk.slot
 import io.mockk.verify
 import no.nav.familie.ef.iverksett.brev.JournalpostClient
 import no.nav.familie.ef.iverksett.brev.domain.KarakterutskriftBrev
-import no.nav.familie.ef.iverksett.oppgave.OppgaveUtil.opprettBrev
+import no.nav.familie.ef.iverksett.brev.frittstående.KarakterInnhentingBrevUtil.opprettBrev
 import no.nav.familie.ef.iverksett.repository.findByIdOrThrow
 import no.nav.familie.kontrakter.ef.felles.FrittståendeBrevType
 import no.nav.familie.kontrakter.felles.dokarkiv.ArkiverDokumentResponse
